// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/replace-abwab-door-link-words-body.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AbwabDoorLinkSelectedWordBody } from '../models/abwab-door-link-selected-word-body';
export interface ReplaceAbwabDoorLinkWordsBody {
  expectedDoorVersion: number | null;
  selectedWords: Array<AbwabDoorLinkSelectedWordBody> | null;
}